requirements.txt
streamlit
openai
python-dotenv
Pillow
requests